requirements.txt
pandas==1.1.5
psycopg2==2.9.3
python-decouple==3.6
python-slugify==6.1.2
requests==2.27.1
sqlalchemy==1.4.40
